app/requirements.txt
python-dotenv
azure-monitor-opentelemetry
